Reject all messages from not upgraded nodes after the hardfork
Upon receiving version message from another node, it checks that if the
hardfork is already activated, then major version of another node version should
not be lower than 5 (checcked using CLIENT_VERSION_MAJOR). If the node
has already sent version message, then check its version before
handling messages, in case if it's not upgraded if the hardfork is
already activated.
This commit also adds additional check for a header version in
CheckBlockHeader, failing if the header has smart-contract hardfork bit
not set after the hardfork. It also fails if hashStateRoot or hashUTXOroot
is null after the hardfork.