Net - NotifyBlockTip signal: switch from hash (uint256) to CBlockIndex * - also add a boolean for specifying if update tip has occurred during initial sync
Changing the logic download to allow getheaders to call / getdata on inbound peers, sent to SendMessages will now call getheaders on both inbound and outbound when the title string is nearly synchronized. It will also try to download blocks from local co-workers when we do not download the original block.