Improve - Randomization when pushing the address in the CNode, round up the fee, or randomly to select the coin.
It replaces its use of insecure_random in the core code with an automatically seeded Seed_Insecure_Rand, and it means that after randomly selecting a maximum number of outputs it aborts. Therefore, the best subsequences never have too many inputs to match in a transaction. This means that is used for inner loops. The Seed_Insecure_Rand can be in the outer scope, or itself, then insecure_rand32 () is used inside the loop. It's useful to push the address into the CNode or to round the fee, or randomly to choose the coin. As a Seed_Insecure_Rand is created for every purpose, therefore it is eliminated from the general use of the non-cross-protected of a set of spheres, which should also eliminate the potential race conditions.