Fix "Don't use change address" feature
Normally, when we send a fund to an address, the wallet will
send the change to newly generated address. The drawback is
that it did not reuse a previous key and the change may be
lost if a backup is restored and we doesn't have the
new private key for the change.

With this fix, it will resolve this issue. Please enable this
feature at option menu "Don't use change address"

Signed-off-by: tran <nguyenhoangtran11@yahoo.com>

(cherry picked from commit ec0228be4027fe62c285e755d482ec1627a23a8c)